+++ CHANGELOG.markdown
@@ -13,6 +13,9 @@
    delegate method and a ViewDidScrollNotification notification that are all
    fired when the view scroll offset changes.
 
+ * [Aphid.UI.View] Added support for handling touch "start", "end" and "move"
+   events.
+
  * [Aphid.UI] Removed all instances of LoadingIndicator, replacing them with
    the new View-based LoadingIndicatorView. This includes moving the global
    loading indicator to Aphid.UI.Window from Aphid.Core.Application.